integrated views
