+++ Pipfile
@@ -8,6 +8,7 @@ django = "*"
 "psycopg2" = "*"
 django-registration-redux = "*"
 djangorestframework = "*"
+django-extensions = "*"
 
 [dev-packages]
 "flake8" = "*"
