+++ api/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path, include
-from questions import views as api_views
+from api import views as api_views
 
 
 urlpatterns = [
